Changed Game dependency in GameObject to GameContext.

+++ game_core/include/game/elixirManager.hpp
@@ -20,6 +20,6 @@ namespace royale
 		void generateElixir(double dt);
 
 		double getElixir() const;
-		void spendElixir(int amount);
+		void spendElixir(double amount);
 	};
 }
